+++ Splito/UI/Home/Expense/Expense Split Option/ExpenseSplitOptionsView.swift
@@ -36,7 +36,7 @@ struct ExpenseSplitOptionsView: View {
                                 SplitOptionsTopView(title: "Split by share", subtitle: "Great for time-based splitting (2 nights -> 2 shares) and splitting across families (family of 3 -> 3 shares).")
                             }
 
-                            SplitOptionsTabView(viewModel: viewModel)
+                            ExpenseSplitOptionsTabView(viewModel: viewModel)
 
                             VSpacer(80)
                         }
@@ -65,35 +65,6 @@ struct ExpenseSplitOptionsView: View {
                     }
                 }
                 .foregroundStyle(primaryColor)
-            }
-        }
-    }
-}
-
-private struct SplitOptionsTabView: View {
-
-    @ObservedObject var viewModel: ExpenseSplitOptionsViewModel
-
-    var body: some View {
-        VStack(spacing: 30) {
-            HStack(spacing: 0) {
-                ForEach(SplitType.allCases, id: \.self) { type in
-                    Picker("Split Option", selection: $viewModel.selectedTab) {
-                        Image(systemName: type.image)
-                            .tag(type)
-                    }
-                    .pickerStyle(SegmentedPickerStyle())
-                }
-                .padding(.horizontal, 12)
-            }
-
-            switch viewModel.selectedTab {
-            case .equally:
-                EqualShareView(viewModel: viewModel)
-            case .percentage:
-                PercentageView(viewModel: viewModel)
-            case .shares:
-                ShareView(viewModel: viewModel)
             }
         }
     }
@@ -134,186 +105,6 @@ private struct SplitOptionsBottomView: View {
         case .shares:
             BottomInfoCardView(title: "\(String(format: "%.0f", viewModel.totalShares)) total shares")
         }
-    }
-}
-
-private struct EqualShareView: View {
-
-    @ObservedObject var viewModel: ExpenseSplitOptionsViewModel
-
-    var body: some View {
-        VStack(spacing: 12) {
-            ForEach(viewModel.groupMembers, id: \.id) { member in
-                ExpenseMemberCellView(member: member, isSelected: viewModel.checkIsMemberSelected(member.id)) {
-                    viewModel.handleMemberSelection(member.id)
-                }
-            }
-        }
-    }
-}
-
-private struct ExpenseMemberCellView: View {
-
-    let member: AppUser
-    let isSelected: Bool
-
-    let onTap: () -> Void
-
-    var body: some View {
-        VStack(spacing: 12) {
-            HStack(alignment: .center, spacing: 20) {
-                MemberProfileImageView(imageUrl: member.imageUrl)
-
-                Text(member.fullName)
-                    .font(.subTitle1())
-                    .foregroundStyle(primaryText)
-
-                Spacer()
-
-                if isSelected {
-                    Image(systemName: "checkmark.circle.fill")
-                        .resizable()
-                        .frame(width: 26, height: 26)
-                        .foregroundStyle(successColor)
-                } else {
-                    Circle()
-                        .strokeBorder(outlineColor.opacity(0.3), lineWidth: 1)
-                        .frame(width: 26, height: 26)
-                }
-            }
-            .padding(.horizontal, 16)
-
-            Divider()
-                .frame(height: 1)
-                .background(outlineColor.opacity(0.3))
-        }
-        .onTouchGesture(onTap)
-    }
-}
-
-private struct PercentageView: View {
-
-    @ObservedObject var viewModel: ExpenseSplitOptionsViewModel
-
-    var body: some View {
-        VStack(spacing: 12) {
-            ForEach(viewModel.groupMembers, id: \.id) { member in
-                MemberCellView(
-                    value: Binding(
-                        get: { viewModel.percentages[member.id] ?? 0 },
-                        set: { viewModel.updatePercentage(for: member.id, percentage: $0) }
-                    ),
-                    member: member, suffixText: "%",
-                    totalValue: viewModel.totalPercentage,
-                    expenseAmount: viewModel.totalAmount,
-                    onChange: { percentage in
-                        viewModel.updatePercentage(for: member.id, percentage: percentage)
-                    }
-                )
-            }
-        }
-    }
-}
-
-private struct ShareView: View {
-
-    @ObservedObject var viewModel: ExpenseSplitOptionsViewModel
-
-    var body: some View {
-        VStack(spacing: 12) {
-            ForEach(viewModel.groupMembers, id: \.id) { member in
-                MemberCellView(
-                    value: Binding(
-                        get: { viewModel.shares[member.id] ?? 0 },
-                        set: { viewModel.updateShare(for: member.id, share: $0) }
-                    ),
-                    member: member, suffixText: "share(s)",
-                    totalValue: viewModel.totalShares,
-                    expenseAmount: viewModel.totalAmount,
-                    onChange: { share in
-                        viewModel.updateShare(for: member.id, share: share)
-                    }
-                )
-            }
-        }
-    }
-}
-
-private struct MemberCellView: View {
-
-    @Binding var value: Double
-
-    let member: AppUser
-    let suffixText: String
-    var totalValue: Double
-    var expenseAmount: Double
-
-    let onChange: (Double) -> Void
-
-    @State private var textValue: String
-
-    init(value: Binding<Double>, member: AppUser, suffixText: String, totalValue: Double, expenseAmount: Double, onChange: @escaping (Double) -> Void) {
-        self._value = value
-        self.member = member
-        self.suffixText = suffixText
-        self.totalValue = totalValue
-        self.expenseAmount = expenseAmount
-        self.onChange = onChange
-        self._textValue = State(initialValue: String(format: "%.0f", value.wrappedValue))
-    }
-
-    var body: some View {
-        VStack(spacing: 12) {
-            HStack(spacing: 20) {
-                MemberProfileImageView(imageUrl: member.imageUrl)
-
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(member.fullName)
-                        .font(.subTitle1())
-                        .foregroundStyle(primaryText)
-
-                    let calculatedValue = totalValue == 0 ? 0 : ((expenseAmount) * (Double(value)) / (totalValue))
-                    Text((calculatedValue).formattedCurrency)
-                        .font(.body2())
-                        .foregroundStyle(disableText)
-                }
-                .multilineTextAlignment(.leading)
-
-                Spacer(minLength: 0)
-
-                HStack(spacing: 4) {
-                    TextField("0", text: $textValue, onCommit: {
-                        updateValue(from: textValue)
-                    })
-                    .font(.subTitle1())
-                    .keyboardType(.decimalPad)
-                    .frame(width: 70)
-                    .textFieldStyle(RoundedBorderTextFieldStyle())
-
-                    Text(suffixText)
-                        .font(.body1())
-                        .foregroundStyle(primaryText)
-                }
-                .fixedSize()
-            }
-            .padding(.horizontal, 16)
-
-            Divider()
-                .frame(height: 1)
-                .background(outlineColor.opacity(0.3))
-        }
-        .onChange(of: textValue) { newValue in
-            updateValue(from: newValue)
-        }
-    }
-
-    private func updateValue(from newValue: String) {
-        if let doubleValue = Double(newValue) {
-            value = doubleValue
-        } else {
-            value = 0
-        }
-        onChange(value)
     }
 }
 
@@ -399,7 +190,7 @@ private struct BottomInfoCardView: View {
                         .foregroundColor(primaryText)
 
                     if value != nil {
-                        Text(value!)
+                        Text(value!.localized)
                             .font(.body1())
                             .foregroundColor(secondaryText)
                     }
